requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.7.0
httpx==0.27.0
python-multipart==0.0.9

# LangGraph / LangChain
langgraph==0.2.0
langchain-core==0.3.0
langchain-groq==0.1.9

# Groq
groq==0.9.0

# Database
aiosqlite==0.20.0
sqlalchemy==2.0.30

# Scheduler
apscheduler==3.10.4

# Social Media
instagrapi==2.1.2
playwright==1.44.0
playwright-stealth==2.0.2
fake-useragent==2.2.0

# Media Generation
gtts==2.5.1
pyttsx3==2.90
moviepy==1.0.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
soundfile==0.12.1
kokoro-onnx==0.5.0

# Utils
python-dotenv==1.0.1
redis==5.0.4
celery==5.4.0
